+++ smrt/src/test/java/com/smart/mapper/ProduitMapperTest.java
@@ -5,8 +5,6 @@ import com.smart.entity.Produit;
 import org.junit.jupiter.api.Test;
 import org.mapstruct.factory.Mappers;
 
-import java.math.BigDecimal;
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
@@ -20,8 +18,8 @@ public class ProduitMapperTest {
         entity.setId("1");
         entity.setNom("Test Nom");
         entity.setCategorie("Test Categorie");
-        entity.setPoids(BigDecimal.valueOf(2.5));
-        entity.setPrix(BigDecimal.valueOf(10.0));
+        entity.setPoids(2.5);
+        entity.setPrix(10.0);
 
         ProduitDTO dto = mapper.toDto(entity);
 
@@ -39,7 +37,7 @@ public class ProduitMapperTest {
         dto.setId("1");
         dto.setNom("Test Nom");
         dto.setCategorie("Test Categorie");
-        dto.setPoids(BigDecimal.valueOf(2.5));
+        dto.setPoids(2.5);
         dto.setPrix(10.0);
 
         Produit entity = mapper.toEntity(dto);
